NormalMap Lighting을 적용하기 위해서 메테리얼, 셰이더, 메시 각각 상속받아 NormalMapMaterial, Shader, Mesh 추가. Shader코드 변경. 이제 Player 노말맵을 제대로 만들어볼 차례

+++ jhLightingManager.cpp
@@ -14,10 +14,10 @@ namespace jh
 	void LightingManager::WriteLightDataAtBuffer()
 	{
 		assert(mspLightStructuredBuffer != nullptr && mspLightConstantBuffer != nullptr);
-		mspLightStructuredBuffer->WirteDataAtBuffer(mLights.data(), mLights.size());
+		mspLightStructuredBuffer->WirteDataAtBuffer(mLights.data(), static_cast<UINT>(mLights.size()));
 		NumberOfLight lightNumber;
 		ZeroMemory(&lightNumber, sizeof(NumberOfLight));
-		lightNumber.NumberOfLight = mLights.size();
+		lightNumber.NumberOfLight = static_cast<UINT>(mLights.size());
 		mspLightConstantBuffer->WirteDataAtBuffer(&lightNumber, sizeof(NumberOfLight));
 	}
 	void LightingManager::SetPipeline()
